Made Rule & Rule Type management global
Why you made the change:

I did this so that it would more easily model having rule shared across
across registries within the same code.
